move gaussian dropout where it can be reused
