+++ src/axolotl/monkeypatch/mistral_attn_hijack_flash.py
@@ -13,7 +13,6 @@ from flash_attn.flash_attn_interface import (  # pylint: disable=ungrouped-impor
     flash_attn_varlen_kvpacked_func,
     flash_attn_varlen_qkvpacked_func,
 )
-from torch import nn
 from transformers import MistralConfig
 from transformers.modeling_outputs import BaseModelOutputWithPast
 from transformers.models.mistral.modeling_mistral import (
@@ -24,26 +23,9 @@ from transformers.models.mistral.modeling_mistral import (
 )
 from transformers.models.mistral.modeling_mistral import apply_rotary_pos_emb, repeat_kv
 
-from axolotl.monkeypatch.utils import get_cu_seqlens_from_pos_ids
+from axolotl.monkeypatch.utils import GaussianDropout, get_cu_seqlens_from_pos_ids
 
 LOG = logging.getLogger("axolotl.monkeypatch.mistral")
-
-
-class GaussianDropout(nn.Module):
-    """
-    Module to apply Gaussian Dropout
-    """
-
-    def __init__(self, p=0.5):  # pylint: disable=invalid-name
-        super().__init__()
-        if p <= 0 or p >= 1:
-            raise ValueError("p value should accomplish 0 < p < 1")
-        self.p = p  # pylint: disable=invalid-name
-
-    def forward(self, inputs):
-        stddev = (self.p / (1.0 - self.p)) ** 0.5
-        epsilon = torch.randn_like(inputs) * stddev
-        return inputs * epsilon
 
 
 def replace_mistral_attn_with_flash_attn(
